refactor: screen_picker.pyをgame_screen_picker.pyにリネーム
