+++ src/main.py
@@ -10,7 +10,7 @@ from .analyzers.image_quality_analyzer import ImageQualityAnalyzer
 from .constants.genre_weights import GenreWeights
 from .models.image_metrics import ImageMetrics
 from .models.picker_statistics import PickerStatistics
-from .services.screen_picker import GameScreenPicker
+from .services.game_screen_picker import GameScreenPicker
 from .utils.file_utils import FileUtils
